+++ src/App.js
@@ -2,8 +2,8 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 
 import Header from 'components/Common/Header/index';
 import MainPage from 'pages/MainPage';
-import TransferPage from 'pages/TransferPage';
-import TransferResultPage from 'pages/TransferResultPage';
+import Transfer from 'pages/Transfer';
+import TransferResult from 'pages/TransferResult';
 import BusSchedulePage from 'pages/BusSchedulePage';
 
 function App() {
@@ -14,8 +14,8 @@ function App() {
 					<Header />
 					<Routes>
 						<Route path='/' element={<MainPage />} />
-						<Route path='/train' element={<TransferPage />} />
-						<Route path='/train/time-table' element={<TransferResultPage />} />
+						<Route path='/train' element={<Transfer />} />
+						<Route path='/train/time-table' element={<TransferResult />} />
 						<Route path='/bus' element={<BusSchedulePage />} />
 					</Routes>
 				</Router>
